fix(batch): load batch after run  (#18)
* Fix typos and update report generation

* Add path column to report_list function and update path value in Citros class

* Fix error messages and provide help instructions

* Fix database connection issue and handle exception in batch uploader

* Refactor Batch class to separate simulation_run method

* Refactor batch.run() method and handle ModuleNotFoundError gracefully

+++ citros/batch_uploader.py
@@ -521,6 +521,9 @@ class BatchUploader:
         table_name = f"{self.name}"
 
         connection = citrosDB.connect()
+        if connection is None:
+            self.log.error("No creating connection to database. Aborting.")
+            raise NoConnectionToCITROSDBException
 
         citrosDB.drop_table(connection, schema_name, table_name)
